+++ apps/electron/src/main/modules/mcp-core/apps/mcp-config-importer.ts
@@ -10,7 +10,7 @@ import {
   exists,
 } from "@/main/modules/mcp-core/client/mcp-app-paths";
 import { v4 as uuidv4 } from "uuid";
-import { getSettingsService } from "@/main/modules/settings/settings-service";
+import { getSettingsService } from "@/main/modules/settings/settings.service";
 
 /**
  * Sync server configurations from a provided list of configs
