+++ apps/electron/src/main/modules/settings/settings.service.ts
